Reconcile move-item and move-dropped-item
