+++ src/app/inventory/StoreBucketDropTarget.tsx
@@ -10,7 +10,7 @@ import {
 } from 'react-dnd';
 import { InventoryBucket } from './inventory-buckets';
 import { DimItem } from './item-types';
-import moveDroppedItem from './move-dropped-item';
+import { moveItemTo } from './move-item';
 import { DimStore } from './store-types';
 
 interface ExternalProps {
@@ -42,7 +42,7 @@ const dropSpec: DropTargetSpec<Props> = {
     // https://github.com/react-dnd/react-dnd-html5-backend/issues/23
     const shiftPressed = (component as StoreBucketDropTarget).shiftKeyDown;
     const item = monitor.getItem().item as DimItem;
-    moveDroppedItem(props.store, item, Boolean(props.equip), shiftPressed);
+    moveItemTo(item, props.store, Boolean(props.equip), item.amount, shiftPressed);
   },
   canDrop(props, monitor) {
     // You can drop anything that can be transferred into a non-equipped bucket
