+++ src/app/inventory/move-item.ts
@@ -1,7 +1,9 @@
+import { DimError } from 'app/bungie-api/bungie-service-helper';
 import { t } from 'app/i18next-t';
 import { hideItemPopup } from 'app/item-popup/item-popup';
 import { PlatformErrorCodes } from 'bungie-api-ts/common';
 import _ from 'lodash';
+import { Subject } from 'rxjs';
 import { showNotification } from '../notifications/notifications';
 import { loadingTracker } from '../shell/loading-tracker';
 import rxStore from '../store/store';
@@ -12,18 +14,97 @@ import { moveItemTo as moveTo } from './item-move-service';
 import { DimItem } from './item-types';
 import { moveItemNotification } from './MoveNotifications';
 import { DimStore } from './store-types';
-import { getVault } from './stores-helpers';
+import { getStore, getVault } from './stores-helpers';
+
+export interface MoveAmountPopupOptions {
+  item: DimItem;
+  targetStore: DimStore;
+  amount: number;
+  maximum: number;
+  onAmountSelected(amount: number);
+  onCancel(): void;
+}
+
+export const showMoveAmountPopup$ = new Subject<MoveAmountPopupOptions>();
+
+function showMoveAmountPopup(
+  item: DimItem,
+  targetStore: DimStore,
+  maximum: number
+): Promise<number> {
+  return new Promise((resolve, reject) => {
+    showMoveAmountPopup$.next({
+      item,
+      targetStore,
+      amount: item.amount,
+      maximum,
+      onAmountSelected: resolve,
+      onCancel: reject,
+    });
+  });
+}
 
 /**
  * Move the item to the specified store. Equip it if equip is true.
  */
 export const moveItemTo = queuedAction(
   loadingTracker.trackPromise(
-    async (item: DimItem, store: DimStore, equip: boolean, amount: number) => {
+    async (
+      item: DimItem,
+      store: DimStore,
+      equip = false,
+      amount: number = item.amount,
+      chooseAmount = false
+    ) => {
       hideItemPopup();
+      if (item.notransfer && item.owner !== store.id) {
+        throw new Error(t('Help.CannotMove'));
+      }
+
+      if (item.owner === store.id && !item.location.inPostmaster) {
+        if ((item.equipped && equip) || (!item.equipped && !equip)) {
+          return;
+        }
+      }
+
+      let moveAmount = amount || 1;
       const reload = item.equipped || equip;
       try {
-        const movePromise = moveTo(item, store, equip, amount);
+        const stores = item.getStoresService().getStores();
+
+        // Select how much of a stack to move
+        if (
+          chooseAmount &&
+          item.maxStackSize > 1 &&
+          item.amount > 1 &&
+          // https://github.com/DestinyItemManager/DIM/issues/3373
+          !item.uniqueStack
+        ) {
+          const maximum = getStore(stores, item.owner)!.amountOfItem(item);
+
+          try {
+            moveAmount = await showMoveAmountPopup(item, store, maximum);
+          } catch (e) {
+            const error: DimError = new Error('move-canceled');
+            error.code = 'move-canceled';
+            throw error;
+          }
+        }
+
+        if ($featureFlags.debugMoves) {
+          console.log(
+            'User initiated move:',
+            moveAmount,
+            item.name,
+            item.type,
+            'to',
+            store.name,
+            'from',
+            getStore(stores, item.owner)!.name
+          );
+        }
+
+        const movePromise = moveTo(item, store, equip, moveAmount);
         showNotification(moveItemNotification(item, store, movePromise));
 
         item = await movePromise;
@@ -35,6 +116,8 @@ export const moveItemTo = queuedAction(
         }
 
         item.updateManualMoveTimestamp();
+
+        return item;
       } catch (e) {
         console.error('error moving item', item.name, 'to', store.name, e);
         // Some errors aren't worth reporting
